refactor: CopyToClipboard with internal pattern (#2248)
